+++ engine/src/main/java/org/terasology/rendering/world/WorldRendererLwjgl.java
@@ -65,7 +65,7 @@ import org.terasology.world.block.Block;
 import org.terasology.world.chunks.ChunkConstants;
 import org.terasology.world.chunks.ChunkProvider;
 import org.terasology.world.chunks.RenderableChunk;
-import org.terasology.world.sun.DefaultCelestialSystem;
+import org.terasology.world.sun.BasicCelestialModel;
 
 import javax.vecmath.Matrix4f;
 import javax.vecmath.Vector3f;
